pyproject.toml
[project]
name = "build-with-syanpse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "pyyaml>=6.0.3",
    "websockets>=15.0.1",
]
